third_party/WebKit/LayoutTests/fast/dom/navigator-maxTouchPoints.html

<html>
<script src="../../resources/js-test.js"></script>
<script>
description('This tests that the maxTouchPoints setting is plumbed through to the navigator object.');

internals.settings.setMaxTouchPoints(37);
shouldBe('navigator.maxTouchPoints', '37');
</script>
<body>
</body>
</html>
