third_party/blink/web_tests/fast/dom/navigator-maxTouchPoints-expected.txt

This tests that the maxTouchPoints setting is plumbed through to the navigator object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.maxTouchPoints is 37
PASS successfullyParsed is true

TEST COMPLETE
